Fix tests and relocate form
